% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{fitting}
\alias{fitting}
\title{Fit a model to the simulated data}
\usage{
fitting(
  sim_result,
  k = 20,
  method = "REML",
  family = gaussian(),
  fit = "gam",
  model = "SiZR"
)
}
\arguments{
\item{sim_result}{A list containing the simulated data frame and the coefficient function f_X, returned from make_data()}

\item{k}{Number of basis functions to use in the smooth term. Default is 20.}

\item{method}{Method for smoothing parameter estimation. Default is "REML".}

\item{family}{Family object specifying the error distribution and link function. Default is gaussian().}

\item{fit}{Type of fitting function to use. Options are "gam" (default) or "bam".}

\item{model}{Type of model to fit. Options are "SiZR" (default) or "mean".}
}
\value{
A fitted model object.
}
\description{
Fit a model to the simulated data
}
\examples{
# Simulate data
sim_data <- sim_data(N = 100, J = 5, s2_e = 1, seed = 123)
# Fit SiZR model
fit_SiZR <- fitting(sim_data, model = "SiZR")
}
