% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{sim_data}
\alias{sim_data}
\title{Simulate data for SiZR model}
\usage{
sim_data(
  N = integer(),
  J = integer(),
  s2_e = numeric(),
  seed = integer(),
  p_X = function(j) runif(j, 0, 1),
  f_X = function(x) x,
  family = c("gaussian", "binomial", "poisson"),
  method = "SiZR"
)
}
\arguments{
\item{N}{Sample size (number of subjects).}

\item{J}{Number of covariates.}

\item{s2_e}{Variance of the residual error term.}

\item{seed}{Random seed for reproducibility.}

\item{p_X}{Function to generate the covariates X. Default is uniform distribution on (0, 1).}

\item{f_X}{Coefficient function defining the association between X and Y. Default is the identity function.}

\item{family}{Family of the outcome variable. Options are "gaussian", "binomial", or "poisson". Default is "gaussian".}

\item{method}{Determines method of data generation. Options are "SiZR" (default) and "mean"}
}
\value{
A list containing: data frame with simulated data and the coefficient function f_X.
}
\description{
Simulate data for SiZR model
}
\examples{
# Simulate data with default parameters
sim_data <- sim_data(N = 100, J = 5, s2_e = 1, seed = 123)
head(sim_data$df)

}
