% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{make_data}
\alias{make_data}
\title{Wrangle data for fitting in SiZR model}
\usage{
make_data(
  data,
  shape = c("long", "wide"),
  predictor = c("functional", "non-functional"),
  Lmat = NULL
)
}
\arguments{
\item{data}{A data frame containing the outcome variable and covariates.}

\item{shape}{A character string specifying the shape of the data. Options are "long" or "wide".}

\item{predictor}{A character string specifying the type of predictor. Options are "functional" or "non-functional".}

\item{Lmat}{An optional matrix of weights for the functional predictor. If NULL, equal weights will be used.}
}
\value{
A data frame formatted for SiZR model fitting.
}
\description{
Wrangle data for fitting in SiZR model
}
